api/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
idna==2.10
install==1.3.4
itypes==1.2.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytrends==4.7.3
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
